+++ configuration.py
@@ -1,4 +1,4 @@
-from constants import CONFIGURATION_FILE, CONFIGURATION_RUN_AUTOLIGHT, CONFIGURATION_RUN_BLINK_TIME_COLON, CONFIGURATION_RUN_CLOCK_TYPE, CONFIGURATION_MQTT_BROKER, CONFIGURATION_MQTT_CONFIG, CONFIGURATION_MQTT_ENABLED, CONFIGURATION_MQTT_PREFIX, CONFIGURATION_RUN_CONFIG, CONFIGURATION_RUN_TEMP, CONFIGURATION_WIFI_CONFIG, CONFIGURATION_WIFI_ENABLED, CONFIGURATION_WIFI_PASSPHRASE, CONFIGURATION_WIFI_SSID
+from constants import CONFIGURATION_FILE, CONFIGURATION_RUN_BLINK_TIME_COLON, CONFIGURATION_RUN_CLOCK_TYPE, CONFIGURATION_RUN_AUTOLIGHT, CONFIGURATION_RUN_SHOW_TEMP, CONFIGURATION_MQTT_BROKER, CONFIGURATION_MQTT_CONFIG, CONFIGURATION_MQTT_ENABLED, CONFIGURATION_MQTT_PREFIX, CONFIGURATION_RUN_CONFIG, CONFIGURATION_RUN_TEMP, CONFIGURATION_WIFI_CONFIG, CONFIGURATION_WIFI_ENABLED, CONFIGURATION_WIFI_HOSTNAME, CONFIGURATION_WIFI_SSID, CONFIGURATION_WIFI_PASSPHRASE, CONFIGURATION_WIFI_NTP_ENABLED, CONFIGURATION_WIFI_NTP_PTZ
 from util import singleton
 from helpers import read_json_file, write_json_file
 
@@ -6,10 +6,13 @@ from helpers import read_json_file, write_json_file
 @singleton
 class Configuration:
     class WifiConfiguration:
-        def __init__(self, enabled: bool, ssid: str, passphrase: str) -> None:
+        def __init__(self, enabled: bool, hostname: str, ssid: str, passphrase: str, ntp_enabled: bool, ntp_ptz: str) -> None:
             self.enabled = enabled
+            self.hostname = hostname
             self.ssid = ssid
             self.passphrase = passphrase
+            self.ntp_enabled = ntp_enabled
+            self.ntp_ptz = ntp_ptz
 
     class MQTTConfiguration:
         def __init__(self, enabled: bool, broker: str, prefix: str) -> None:
@@ -23,6 +26,7 @@ class Configuration:
         self.blink_time_colon = False
         self.temp = "c"
         self.clock_type = "24"
+        self.show_temp = True
         self.autolight = False
         self.read_config_file()
 
@@ -34,12 +38,16 @@ class Configuration:
         self.blink_time_colon = self.config[CONFIGURATION_RUN_CONFIG][CONFIGURATION_RUN_BLINK_TIME_COLON]
         self.temp = self.config[CONFIGURATION_RUN_CONFIG][CONFIGURATION_RUN_TEMP]
         self.clock_type = self.config[CONFIGURATION_RUN_CONFIG][CONFIGURATION_RUN_CLOCK_TYPE]
+        self.show_temp = self.config[CONFIGURATION_RUN_CONFIG][CONFIGURATION_RUN_SHOW_TEMP]
         self.autolight = self.config[CONFIGURATION_RUN_CONFIG][CONFIGURATION_RUN_AUTOLIGHT]
 
         self.wifi_config = self.WifiConfiguration(
             enabled=self.config[CONFIGURATION_WIFI_CONFIG][CONFIGURATION_WIFI_ENABLED],
+            hostname=self.config[CONFIGURATION_WIFI_CONFIG][CONFIGURATION_WIFI_HOSTNAME],
             ssid=self.config[CONFIGURATION_WIFI_CONFIG][CONFIGURATION_WIFI_SSID],
-            passphrase=self.config[CONFIGURATION_WIFI_CONFIG][CONFIGURATION_WIFI_PASSPHRASE]
+            passphrase=self.config[CONFIGURATION_WIFI_CONFIG][CONFIGURATION_WIFI_PASSPHRASE],
+            ntp_enabled=self.config[CONFIGURATION_WIFI_CONFIG][CONFIGURATION_WIFI_NTP_ENABLED],
+            ntp_ptz=self.config[CONFIGURATION_WIFI_CONFIG][CONFIGURATION_WIFI_NTP_PTZ]
         )
 
         self.mqtt_config = self.MQTTConfiguration(